refactor(user-skills): extract personal skill store
