+++ scripts/ci/module-impact.json
@@ -215,6 +215,7 @@
     "user_skills_repository": {
       "ownerPaths": [
         "src/main/skills/user-skill-repository.ts",
+        "src/main/skills/user-skill-store.ts",
         "src/main/skills/skill-mutation-owner.ts",
         "src/main/skills/skill-package-transaction-owner.ts"
       ],
