+++ scripts/ci/module-test-impact.test.ts
@@ -70,6 +70,7 @@ describe('module test impact commands', () => {
 
     for (const path of [
       'src/main/skills/user-skill-repository.ts',
+      'src/main/skills/user-skill-store.ts',
       'src/main/skills/skill-mutation-owner.ts',
       'src/main/skills/skill-package-transaction-owner.ts'
     ]) {
